+++ src/index.ts
@@ -1,2 +1,4 @@
 export { connect, RxReactContainer } from './connect';
 export { combineProps } from './combineProps';
+export { useRxController } from './useRxController';
+export { PropsHelper } from './PropsHelper';