+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.9.0 (2019-08-28)
+
+- `useRxContainer` hook, as alternative to `connect` function
+- *breaking* restricting what can be accessed from `controller`, limiting it to only `props` (instead of wrapper component instance) 
+
 ## 0.8.0 (2019-08-27)
 
 - better type coverage